Add files via upload
Task-01 Text Generation with GPT-2: Train a model to generate coherent and contextually relevant text based on a given prompt. Starting with GPT- 2, a transformer model developed by Open Al, you will learn how to fine-tune the model on a custom dataset to create text that mimics the style and structure of your training data.